refactor(card): rename selectCardsForDeck to logic
