+++ src/entities/card/model/logic.ts
